move address fns to address.go, message fns to message.go
