+++ dcrlibwallet.go
@@ -8,17 +8,13 @@ import (
 	"path/filepath"
 	"strings"
 	"syscall"
-	"time"
 
 	"github.com/asdine/storm"
-	"github.com/decred/dcrd/dcrec"
 	"github.com/decred/dcrd/dcrjson"
-	"github.com/decred/dcrd/dcrutil"
 	"github.com/decred/dcrwallet/errors"
 	"github.com/decred/dcrwallet/netparams"
 	"github.com/decred/dcrwallet/wallet"
 	"github.com/decred/dcrwallet/wallet/txrules"
-	"github.com/raedahgroup/dcrlibwallet/addresshelper"
 	"github.com/raedahgroup/dcrlibwallet/utils"
 	"go.etcd.io/bbolt"
 )
@@ -150,73 +146,6 @@ func (lw *LibWallet) Shutdown(exit bool) {
 	if exit {
 		os.Exit(0)
 	}
-}
-
-func (lw *LibWallet) SignMessage(passphrase []byte, address string, message string) ([]byte, error) {
-	lock := make(chan time.Time, 1)
-	defer func() {
-		lock <- time.Time{}
-	}()
-	err := lw.wallet.Unlock(passphrase, lock)
-	if err != nil {
-		return nil, translateError(err)
-	}
-
-	addr, err := addresshelper.DecodeForNetwork(address, lw.wallet.ChainParams())
-	if err != nil {
-		return nil, translateError(err)
-	}
-
-	var sig []byte
-	switch a := addr.(type) {
-	case *dcrutil.AddressSecpPubKey:
-	case *dcrutil.AddressPubKeyHash:
-		if a.DSA(a.Net()) != dcrec.STEcdsaSecp256k1 {
-			return nil, errors.New(ErrInvalidAddress)
-		}
-	default:
-		return nil, errors.New(ErrInvalidAddress)
-	}
-
-	sig, err = lw.wallet.SignMessage(message, addr)
-	if err != nil {
-		return nil, translateError(err)
-	}
-
-	return sig, nil
-}
-
-func (lw *LibWallet) VerifyMessage(address string, message string, signatureBase64 string) (bool, error) {
-	var valid bool
-
-	addr, err := dcrutil.DecodeAddress(address)
-	if err != nil {
-		return false, translateError(err)
-	}
-
-	signature, err := DecodeBase64(signatureBase64)
-	if err != nil {
-		return false, err
-	}
-
-	// Addresses must have an associated secp256k1 private key and therefore
-	// must be P2PK or P2PKH (P2SH is not allowed).
-	switch a := addr.(type) {
-	case *dcrutil.AddressSecpPubKey:
-	case *dcrutil.AddressPubKeyHash:
-		if a.DSA(a.Net()) != dcrec.STEcdsaSecp256k1 {
-			return false, errors.New(ErrInvalidAddress)
-		}
-	default:
-		return false, errors.New(ErrInvalidAddress)
-	}
-
-	valid, err = wallet.VerifyMessage(message, addr, signature)
-	if err != nil {
-		return false, translateError(err)
-	}
-
-	return valid, nil
 }
 
 func (lw *LibWallet) CallJSONRPC(method string, args string, address string, username string, password string, caCert string) (string, error) {
